+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This page reports the main releases only and the main changes therein.
 
 ## [v1.11.x] - 2025-12-10
 
+### Added
+- startup dialog reporting new available version
+
 ### Fixed
 - fixed version retrival if ssl fails in about dialog